requirements.txt
werkzeug==3.1
Flask==3.1.0
flask_sqlalchemy==3.1.1
flask_bcrypt==1.0.1
flask_jwt_extended==4.7.1
